refactor: split utils init in files
